refactor(dealer): isStay 를 의미를 드러내며, 부정 표현을 안쓰도록 shouldDraw로 변경
