--- src/main/java/blackjack/domain/betting/Betting.java
@@ -1,22 +1,0 @@
-package blackjack.domain.betting;
-
-import blackjack.domain.participant.Player;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class Betting {
-    private final Map<Player, Money> bettingTable = new HashMap<>();
-
-    public void put(Player player, Money money) {
-        bettingTable.put(player, money);
-    }
-
-    public Money get(Player player) {
-        return bettingTable.get(player);
-    }
-
-    public double revenue(Player player) {
-        return get(player).value() * player.profitRatio();
-    }
-}
